+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to "Extension Update Reporter" extension will be documented 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## Unreleased
+### Added
+- Add command to show report. Closes [#8](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/8) by @jerone in [PR #38](https://github.com/ryu1kn/vscode-extension-update-reporter/pull/38).
+
 ## [1.2.0] - 2022-11-06
 ### Added
 - Marketplace, extension homepage links are always available for easier access to the full information. (thanks to @jerone [PR #29](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/29))